pyproject.toml
[project]
name = "trion"
version = "0.1.0"
description = "Transdisciplinary Research Intelligence & Opportunity Navigation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Amir", email = "amir@example.com"}]

dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "pyyaml>=6.0",
    "deepagents>=0.3.12",
]

[project.optional-dependencies]
research = [
    "biopython>=1.80",
    "arxiv>=2.0.0",
    "pandas>=2.0.0",
]
document = [
    "docling>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]
web = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "python-jobspy>=1.1.0",
    "ddgs>=0.2.0",
]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
]
all = [
    "trion[research,document,web,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trion"]

[tool.hatch.build.targets.wheel.force-include]
"trion/py.typed" = "trion/py.typed"

[tool.uv]
package = true
